refactor: Refactor Drawer components into dedicated package
- Moved DrawerContent and related composables (DrawerHeader, DrawerMenuItem, PetTypeButton, PetTypeSwitcher, VersionInfo) into a new 'drawer' package
- Updated imports and parameter order for consistency across UI components.
- Minor code cleanups and improvements in modifier usage.
