+++ core/ui/src/main/java/com/example/ui/components/ErrorMessage.kt
@@ -12,9 +12,9 @@ import androidx.compose.ui.unit.dp
 
 @Composable
 fun ErrorMessage(
+    modifier: Modifier = Modifier,
     message: String,
-    onRetry: () -> Unit,
-    modifier: Modifier = Modifier
+    onRetry: () -> Unit
 ) {
     Column(
         modifier = modifier
